http_backoff retry with SliceFileObj
